Convert rating from int to float64 and remove query parameter to get rating of book

+++ internal/bookreviews/application/create/BookReviewCommand.go
@@ -4,17 +4,17 @@ import validation "github.com/go-ozzo/ozzo-validation"
 
 // BookReviewCommand ...
 type BookReviewCommand struct {
-	ID     string `json:"id"`
-	Text   string `json:"text"`
-	Rating int    `json:"rating"`
-	BookID string `json:"book_id"`
-	UserID string `json:"user_id"`
+	ID     string  `json:"id"`
+	Text   string  `json:"text"`
+	Rating float64 `json:"rating"`
+	BookID string  `json:"book_id"`
+	UserID string  `json:"user_id"`
 }
 
 // Validate ...
 func (b BookReviewCommand) Validate() error {
 	return validation.ValidateStruct(&b,
 		validation.Field(&b.Text, validation.Required, validation.Length(1, 250)),
-		validation.Field(&b.Rating, validation.Required, validation.Min(1), validation.Max(5)),
+		validation.Field(&b.Rating, validation.Required, validation.Min(1.0), validation.Max(5.0)),
 	)
 }
