+++ internal/bookreviews/domain/BookReview.go
@@ -6,14 +6,14 @@ import "time"
 type BookReview struct {
 	ID        string
 	Text      string
-	Rating    int
+	Rating    float64
 	BookID    string
 	UserID    string
 	CreatedOn time.Time
 }
 
 // NewBookReview ...
-func NewBookReview(id, text string, rating int, bookID, userID string) (*BookReview, error) {
+func NewBookReview(id, text string, rating float64, bookID, userID string) (*BookReview, error) {
 	return &BookReview{
 		ID:        id,
 		Text:      text,
